requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.0
python-dotenv==1.0.0
